server/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
beautifulsoup4==4.12.2
firebase-admin==6.2.0